backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
librosa
matplotlib
Pillow
numpy
soundfile